Rewrite test interface
