+++ lb-bdd.js
@@ -6,131 +6,33 @@
 'use strict';
 
 var Mocha = require('mocha');
-var Suite = require('mocha/lib/suite');
-var Test = require('mocha/lib/test');
-var escapeRe = require('escape-string-regexp');
+var skip_functions = require('./bdd_skip_functions');
 
-/**
- * BDD-style interface:
- *
- *      describe('Array', function() {
- *        describe('#indexOf()', function() {
- *          it('should return -1 when not present', function() {
- *            // ...
- *          });
- *
- *          it('should return the index when present', function() {
- *            // ...
- *          });
- *        });
- *      });
- *
- * @param {Suite} suite Root suite.
- */
+
 module.exports = Mocha.interfaces['lb-bdd'] = function(suite) {
   var suites = [suite];
 
   var skips = [];
   if (process.env.LOOPBACK_MOCHA_SKIPS) {
-    skips = process.env.LOOPBACK_MOCHA_SKIPS;
+    skips = JSON.parse(process.env.LOOPBACK_MOCHA_SKIPS);
   }
 
   suite.on('pre-require', function(context, file, mocha) {
     var common = require('mocha/lib/interfaces/common')(suites, context);
+    var iface = skip_functions(suites, context, file, skips, Mocha);
 
     context.before = common.before;
     context.after = common.after;
     context.beforeEach = common.beforeEach;
     context.afterEach = common.afterEach;
     context.run = mocha.options.delay && common.runWithSuite(suite);
-    /**
-     * Describe a "suite" with the given `title`
-     * and callback `fn` containing nested suites
-     * and/or tests.
-     */
 
-    context.describe = context.context = function(title, fn) {
-      var pending = false;
-
-      if (skips.indexOf(title) > -1) {
-        pending = true;
-        title = '** SKIPPED BY LB-BDD ** ' + title;
-      }
-
-      var suite = Suite.create(suites[0], title);
-      suite.pending = suite.pending || pending;
-
-      suite.file = file;
-      suites.unshift(suite);
-      fn.call(suite);
-      suites.shift();
-      return suite;
-    };
-
-    /**
-     * Pending describe.
-     */
-
-    context.xdescribe = context.xcontext = context.describe.skip = function(title, fn) {
-      var suite = Suite.create(suites[0], title);
-      suite.pending = true;
-      suites.unshift(suite);
-      fn.call(suite);
-      suites.shift();
-    };
-
-    /**
-     * Exclusive suite.
-     */
-
-    context.describe.only = function(title, fn) {
-      var suite = context.describe(title, fn);
-      mocha.grep(suite.fullTitle());
-      return suite;
-    };
-
-    /**
-     * Describe a specification or test-case
-     * with the given `title` and callback `fn`
-     * acting as a thunk.
-     */
-
-    var it = context.it = context.specify = function(title, fn) {
-      var suite = suites[0];
-      if (suite.isPending() || skips.indexOf(title) > -1) {
-        fn = null;
-        if (skips.indexOf(title) > -1) title = '** SKIPPED BY LB-BDD ** ' + title;
-      }
-      var test = new Test(title, fn);
-      test.file = file;
-      suite.addTest(test);
-      return test;
-    };
-
-    /**
-     * Exclusive test-case.
-     */
-
-    context.it.only = function(title, fn) {
-      var test = it(title, fn);
-      var reString = '^' + escapeRe(test.fullTitle()) + '$';
-      mocha.grep(new RegExp(reString));
-      return test;
-    };
-
-    /**
-     * Pending test case.
-     */
-
-    context.xit = context.xspecify = context.it.skip = function(title) {
-      context.it(title);
-    };
-
-    /**
-     * Number of attempts to retry.
-     */
-    context.it.retries = function(n) {
-      context.retries(n);
-    };
+    context.describe = context.context = iface.describe;
+    context.xdescribe = context.xcontext = context.describe.skip = iface.xdescribe;
+    context.describe.only = iface.describe_only;
+    context.it = context.specify = iface.it;
+    context.it.only = iface.it_only;
+    context.xit = context.xspecify = context.it.skip = iface.xit;
+    context.it.retries = iface.it_retries;
   });
 };
